+++ lib/td_dd_web/controllers/data_structure_version_controller.ex
@@ -42,7 +42,7 @@ defmodule TdDdWeb.DataStructureVersionController do
     :with_confidential,
     :grant,
     :grants,
-    :implementations,
+    :implementation_count,
     :published_note
   ]
 
@@ -149,10 +149,7 @@ defmodule TdDdWeb.DataStructureVersionController do
   end
 
   defp get_data_structure_version(data_structure_id, "latest", opts) do
-    DataStructures.get_latest_version(
-      data_structure_id,
-      opts
-    )
+    DataStructures.get_latest_version(data_structure_id, opts)
   end
 
   defp get_data_structure_version(data_structure_id, version, opts) do