+++ lib/td_dd_web/views/data_structure_version_view.ex
@@ -4,7 +4,6 @@ defmodule TdDdWeb.DataStructureVersionView do
   alias TdDd.DataStructures
   alias TdDdWeb.GrantView
   alias TdDdWeb.StructureTagView
-  alias TdDqWeb.ImplementationStructureView
 
   @protected DataStructures.protected()
 
@@ -39,7 +38,6 @@ defmodule TdDdWeb.DataStructureVersionView do
     |> add_profile()
     |> add_embedded_relations(dsv)
     |> merge_metadata()
-    |> merge_implementations()
     |> add_data_structure_type()
     |> add_note()
     |> add_tags(assigns)
@@ -63,7 +61,7 @@ defmodule TdDdWeb.DataStructureVersionView do
       :grants,
       :group,
       :id,
-      :implementations,
+      :implementation_count,
       :links,
       :data_structure_links,
       :metadata,
@@ -214,7 +212,7 @@ defmodule TdDdWeb.DataStructureVersionView do
     with_profile_attrs(dsv, profile)
   end
 
-  defp add_profile(dsv), do: dsv
+  defp add_profile(dsv), do: Map.delete(dsv, :profile)
 
   defp field_structure_json(
          %{
@@ -249,8 +247,7 @@ defmodule TdDdWeb.DataStructureVersionView do
   end
 
   defp version_json(version) do
-    version
-    |> Map.take([:version, :deleted_at, :inserted_at, :updated_at])
+    Map.take(version, [:version, :deleted_at, :inserted_at, :updated_at])
   end
 
   defp add_system(dsv) do
@@ -332,16 +329,6 @@ defmodule TdDdWeb.DataStructureVersionView do
   end
 
   defp merge_metadata(dsv), do: dsv
-
-  defp merge_implementations(%{implementations: [_ | _] = implementations} = dsv) do
-    Map.put(
-      dsv,
-      :implementations,
-      render_many(implementations, ImplementationStructureView, "implementation_structure.json")
-    )
-  end
-
-  defp merge_implementations(dsv), do: dsv
 
   defp add_data_structure_type(%{data_structure_type: nil} = dsv),
     do: Map.put(dsv, :data_structure_type, %{})