+++ lib/td_dq/rules/search/aggregations.ex
@@ -9,6 +9,7 @@ defmodule TdDq.Rules.Search.Aggregations do
   def aggregations do
     static_aggs = %{
       "active.raw" => %{terms: %{field: "active.raw"}},
+      "df_label.raw" => %{terms: %{field: "df_label.raw", size: 50}},
       "taxonomy" => %{terms: %{field: "domain_ids", size: 500}}
     }